+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Added `bind(myFlowFactory) { component -> component.createMyFlowDependencies() }` binding method.
 - Rename `KeyState` to `FragmentState` and move it to `android` package.
 - Move `ActivityStoreContext` into `android` package.
+- Move `ActivityStore` into `android` package.
 
 ## [0.6.1] - November 18, 2020
 - Bugfix: Fix runtime ignoring `Formula.key` for the root formula.